Renames edit and add views
