+++ src/view/add-new-event-view.js
